:recycle: Move repeated test utilities code to helpers file
